+++ pkg/txm/clientwrappers/dualbroadcast/selector_test.go
@@ -32,6 +32,7 @@ func TestSelectClient_FlashbotsPrimaryOnly(t *testing.T) {
 
 	mux, ok := c.(*multiplexClient)
 	require.True(t, ok, "single flashbots URL should still use multiplexClient as the outer shell")
+	assert.Equal(t, "flashbots", mux.primaryBackend)
 	assert.Empty(t, mux.secondaries)
 	fb, ok := mux.primary.(*ofaTXClient)
 	require.True(t, ok)
@@ -54,6 +55,7 @@ func TestSelectClient_FlashbotsPrimaryWithNovaSecondary(t *testing.T) {
 
 	mux, ok := c.(*multiplexClient)
 	require.True(t, ok)
+	assert.Equal(t, "flashbots", mux.primaryBackend)
 	pri, ok := mux.primary.(*ofaTXClient)
 	require.True(t, ok)
 	assert.Equal(t, ofaKindFlashbots, pri.kind)
@@ -75,6 +77,7 @@ func TestSelectClient_NovaPrimaryOnly(t *testing.T) {
 
 	mux, ok := c.(*multiplexClient)
 	require.True(t, ok)
+	assert.Equal(t, "nova", mux.primaryBackend)
 	assert.Empty(t, mux.secondaries)
 	nc, ok := mux.primary.(*ofaTXClient)
 	require.True(t, ok, "nova URL should use OFA primary inside multiplex shell")
@@ -94,9 +97,16 @@ func TestSelectClient_MetaPrimaryOnly(t *testing.T) {
 
 	mux, ok := c.(*multiplexClient)
 	require.True(t, ok)
+	assert.Equal(t, "meta", mux.primaryBackend)
 	assert.Empty(t, mux.secondaries)
 	_, isMeta := mux.primary.(*MetaClient)
 	require.True(t, isMeta)
+}
+
+func TestBackendLabel(t *testing.T) {
+	assert.Equal(t, "flashbots", backendLabel(&ofaTXClient{kind: ofaKindFlashbots}))
+	assert.Equal(t, "nova", backendLabel(&ofaTXClient{kind: ofaKindNova}))
+	assert.Equal(t, "meta", backendLabel(&MetaClient{}))
 }
 
 func TestRedactURL(t *testing.T) {
